Move healthcheck util functions to util.go
