+++ pkg/api/service/annotations.go
@@ -16,13 +16,6 @@ limitations under the License.
 
 package service
 
-import (
-	"strconv"
-
-	"github.com/golang/glog"
-	"k8s.io/kubernetes/pkg/api"
-)
-
 const (
 	// AnnotationLoadBalancerSourceRangesKey is the key of the annotation on a service to set allowed ingress ranges on their LoadBalancers
 	//
@@ -55,45 +48,3 @@ const (
 	// BetaAnnotationExternalTraffic is the beta version of AlphaAnnotationExternalTraffic.
 	BetaAnnotationExternalTraffic = "service.beta.kubernetes.io/external-traffic"
 )
-
-// NeedsHealthCheck Check service for health check annotations
-func NeedsHealthCheck(service *api.Service) bool {
-	// First check the alpha annotation and then the beta. This is so existing
-	// Services continue to work till the user decides to transition to beta.
-	// If they transition to beta, there's no way to go back to alpha without
-	// rolling back the cluster.
-	for _, annotation := range []string{AlphaAnnotationExternalTraffic, BetaAnnotationExternalTraffic} {
-		if l, ok := service.Annotations[annotation]; ok {
-			if l == AnnotationValueExternalTrafficLocal {
-				return true
-			} else if l == AnnotationValueExternalTrafficGlobal {
-				return false
-			} else {
-				glog.Errorf("Invalid value for annotation %v: %v", annotation, l)
-			}
-		}
-	}
-	return false
-}
-
-// GetServiceHealthCheckNodePort Return health check node port annotation for service, if one exists
-func GetServiceHealthCheckNodePort(service *api.Service) int32 {
-	if !NeedsHealthCheck(service) {
-		return 0
-	}
-	// First check the alpha annotation and then the beta. This is so existing
-	// Services continue to work till the user decides to transition to beta.
-	// If they transition to beta, there's no way to go back to alpha without
-	// rolling back the cluster.
-	for _, annotation := range []string{AlphaAnnotationHealthCheckNodePort, BetaAnnotationHealthCheckNodePort} {
-		if l, ok := service.Annotations[annotation]; ok {
-			p, err := strconv.Atoi(l)
-			if err != nil {
-				glog.Errorf("Failed to parse annotation %v: %v", annotation, err)
-				continue
-			}
-			return int32(p)
-		}
-	}
-	return 0
-}
